+++ src/gluonts/model/seq2seq/_seq2seq_estimator.py
@@ -36,7 +36,7 @@ from gluonts.model.estimator import GluonEstimator
 from gluonts.model.forecast import QuantileForecast, Quantile
 from gluonts.model.predictor import Predictor, RepresentableBlockPredictor
 from gluonts.support.util import copy_parameters
-from gluonts.time_feature.lag import time_features_from_frequency_str
+from gluonts.time_feature import time_features_from_frequency_str
 from gluonts.trainer import Trainer
 from gluonts.transform import ExpectedNumInstanceSampler
