+++ src/gluonts/model/transformer/_estimator.py
@@ -24,7 +24,7 @@ from gluonts.distribution import StudentTOutput, DistributionOutput
 from gluonts.model.estimator import GluonEstimator
 from gluonts.model.predictor import Predictor, RepresentableBlockPredictor
 from gluonts.support.util import copy_parameters
-from gluonts.time_feature.lag import (
+from gluonts.time_feature import (
     TimeFeature,
     get_lags_for_frequency,
     time_features_from_frequency_str,
