+++ src/gluonts/model/wavenet/_estimator.py
@@ -33,7 +33,7 @@ from gluonts.support.util import (
     copy_parameters,
     get_hybrid_forward_input_names,
 )
-from gluonts.time_feature.lag import time_features_from_frequency_str
+from gluonts.time_feature import time_features_from_frequency_str
 from gluonts.trainer import Trainer
 from gluonts.transform import (
     AddAgeFeature,
